third_party/WebKit/LayoutTests/external/wpt/html/browsers/browsing-the-web/navigating-across-documents/015-expected.txt

This is a testharness.js-based test.
FAIL  Link with javascript onclick and href script order  assert_array_equals: property 1, expected "after script" but got "href"
Harness: the test ran to completion.
